requirements.txt
# Requirements for the project
streamlit
requests
beautifulsoup4
parse
python-docx
openai